+++ src/puripuly_heart/ui/app.py
@@ -145,6 +145,9 @@ class TranslatorApp:
         self.view_dashboard.on_enroll_speaker = (
             lambda cluster_id, name: self.controller.enroll_speaker(cluster_id, name)
         )
+        self.view_dashboard.on_speaker_name_lookup = (
+            lambda cluster_id: self.controller.speaker_name_for_cluster(cluster_id)
+        )
         self.view_dashboard.on_toggle_translation = self._on_translation_toggle
         self.view_dashboard.on_toggle_stt = self._on_stt_toggle
         self.view_dashboard.on_toggle_overlay = self._on_overlay_toggle